添加发送语音功能以及发送截图
