+++ package.json
@@ -111,6 +111,7 @@
     "electron-builder": "^20.38.1",
     "electron-debug": "^1.5.0",
     "electron-devtools-installer": "^2.2.4",
+    "electron-log": "^3.0.6",
     "electron-updater": "^4.0.5",
     "eslint": "^4.19.1",
     "eslint-config-standard": "^11.0.0",
@@ -149,7 +150,6 @@
     "vue-loader": "^15.2.4",
     "vue-style-loader": "^4.1.0",
     "vue-template-compiler": "^2.6.6",
-    "vue-worker": "^1.2.1",
     "webpack": "^4.15.1",
     "webpack-cli": "^3.0.8",
     "webpack-dev-server": "^3.1.4",
